加入了验证数据验证分数

+++ src/main/java/com/westlake/air/pecs/domain/db/AnalyseOverviewDO.java
@@ -38,9 +38,9 @@ public class AnalyseOverviewDO extends BaseDO {
 
     String iRtLibraryName;
 
-    Float slope;
+    Double slope;
 
-    Float intercept;
+    Double intercept;
 
     //MS1是全光谱扫描的.没有RTwindow的参数,MS2有
     Float rtExtractWindow;
